New feature: Ignore tags for TurnOnMachines
It's possible to ignore tags for the TurnOnMachines-Action by adding them to the configuration. This new option defaults to the "LargeDoor"-tag which is the cause for this new option.

Added as option to ignore more tags if needed as the worldobjects could have an OnOffComponent and need to be ignored also -> Used tags as those are the most global filter that could be applied (and the mod itself is done by a component which can be included by many worldobjects)